requirements.txt
adal==1.2.7
azure-common==1.1.4
azure-core==1.21.1
azure-identity==1.7.1
azure-mgmt-core==1.3.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==0.31.0
azure-mgmt-sql==0.2.0
azure-mgmt-storage==19.0.0
azure-nspkg==3.0.2
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
fake-useragent==0.1.11
humanize==3.13.1
idna==3.3
isodate==0.6.1
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
oauthlib==3.1.1
portalocker==2.3.2
pycparser==2.21
PyJWT==2.3.0
pyodbc==4.0.32
python-dateutil==2.8.2
python-dotenv==0.19.2
pywin32==303
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
tqdm==4.62.3
urllib3==1.26.8
wget==3.2
